requirements.txt
Django==1.7.4
django-allauth==0.19.1
django-cors-headers==1.0.0
django-debug-toolbar==1.2.2
django-filter==0.9.2
django-rest-auth==0.3.4
django-rest-swagger==0.2.8
djangorestframework==3.0.5
Markdown==2.6
oauthlib==0.7.2
python-openid==2.2.5
PyYAML==3.11
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.1.14
